pyproject.toml
[project]
name = "plaid-mcp-server"
version = "1.0.0"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "mcp[cli]>=1.0.0",
    "plaid-python>=9.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
plaid-mcp-server = "src.server:main"
